+++ spec/classes/install_spec.rb
@@ -24,27 +24,27 @@ describe 'crowd::install' do
         it { is_expected.to contain_group('crowd') }
 
         it { is_expected.to contain_file('/opt/crowd') }
-        it { is_expected.to contain_file('/opt/crowd/atlassian-crowd-2.11.1-standalone') }
+        it { is_expected.to contain_file('/opt/crowd/atlassian-crowd-3.0.1-standalone') }
         it { is_expected.to contain_file('/var/local/crowd') }
 
-        it { is_expected.to contain_staging__file('atlassian-crowd-2.11.1.tar.gz').with({
-          :source => 'https://www.atlassian.com/software/crowd/downloads/binary/atlassian-crowd-2.11.1.tar.gz',
+        it { is_expected.to contain_staging__file('atlassian-crowd-3.0.1.tar.gz').with({
+          :source => 'https://www.atlassian.com/software/crowd/downloads/binary/atlassian-crowd-3.0.1.tar.gz',
         })}
 
-        it { is_expected.to contain_staging__extract('atlassian-crowd-2.11.1.tar.gz').with({
-          :target  => '/opt/crowd/atlassian-crowd-2.11.1-standalone',
-          :creates => '/opt/crowd/atlassian-crowd-2.11.1-standalone/apache-tomcat',
+        it { is_expected.to contain_staging__extract('atlassian-crowd-3.0.1.tar.gz').with({
+          :target  => '/opt/crowd/atlassian-crowd-3.0.1-standalone',
+          :creates => '/opt/crowd/atlassian-crowd-3.0.1-standalone/apache-tomcat',
           :user    => 'crowd',
           :group   => 'crowd',
         })}
 
         it { is_expected.to contain_staging__file('jdbc driver').with({
           :source => 'http://central.maven.org/maven2/mysql/mysql-connector-java/5.1.36/mysql-connector-java-5.1.36.jar',
-          :target => '/opt/crowd/atlassian-crowd-2.11.1-standalone/apache-tomcat/lib/mysql-connector-java-5.1.36.jar',
+          :target => '/opt/crowd/atlassian-crowd-3.0.1-standalone/apache-tomcat/lib/mysql-connector-java-5.1.36.jar',
         })}
 
-        it { is_expected.to contain_exec('chown_/opt/crowd/atlassian-crowd-2.11.1-standalone').with({
-          :command => 'chown -R crowd:crowd /opt/crowd/atlassian-crowd-2.11.1-standalone',
+        it { is_expected.to contain_exec('chown_/opt/crowd/atlassian-crowd-3.0.1-standalone').with({
+          :command => 'chown -R crowd:crowd /opt/crowd/atlassian-crowd-3.0.1-standalone',
         })}
       end
     end
@@ -52,13 +52,13 @@ describe 'crowd::install' do
     context 'update crowd' do
       let(:facts) do
         facts.merge({
-          :crowd_version => '2.9.0'
+          :crowd_version => '2.11.1'
         })
       end
       let :pre_condition do
-        "class { 'crowd': version => '2.11.1' }"
+        "class { 'crowd': version => '3.0.1' }"
       end
-      it { is_expected.to contain_notify('Updating Crowd from version 2.9.0 to 2.11.1') }
+      it { is_expected.to contain_notify('Updating Crowd from version 2.11.1 to 3.0.1') }
       it { is_expected.to contain_exec('stop crowd for update').with({
         :command => 'service crowd stop && sleep 10',
       })}
@@ -72,10 +72,10 @@ describe 'crowd::install' do
           "class { 'crowd': installdir => '/svr/crowd' }"
         end
         it { is_expected.to contain_file('/svr/crowd') }
-        it { is_expected.to contain_file('/svr/crowd/atlassian-crowd-2.11.1-standalone') }
-        it { is_expected.to contain_staging__extract('atlassian-crowd-2.11.1.tar.gz').with({
-          :target => '/svr/crowd/atlassian-crowd-2.11.1-standalone',
-          :creates => '/svr/crowd/atlassian-crowd-2.11.1-standalone/apache-tomcat',
+        it { is_expected.to contain_file('/svr/crowd/atlassian-crowd-3.0.1-standalone') }
+        it { is_expected.to contain_staging__extract('atlassian-crowd-3.0.1.tar.gz').with({
+          :target => '/svr/crowd/atlassian-crowd-3.0.1-standalone',
+          :creates => '/svr/crowd/atlassian-crowd-3.0.1-standalone/apache-tomcat',
         })}
       end
 
@@ -88,17 +88,17 @@ describe 'crowd::install' do
 
       describe 'version' do
         let :pre_condition do
-          "class { 'crowd': version => '2.11.1' }"
+          "class { 'crowd': version => '3.0.1' }"
         end
-        it { is_expected.to contain_file('/opt/crowd/atlassian-crowd-2.11.1-standalone') }
+        it { is_expected.to contain_file('/opt/crowd/atlassian-crowd-3.0.1-standalone') }
 
-        it { is_expected.to contain_staging__file('atlassian-crowd-2.11.1.tar.gz').with({
-          :source => 'https://www.atlassian.com/software/crowd/downloads/binary/atlassian-crowd-2.11.1.tar.gz',
+        it { is_expected.to contain_staging__file('atlassian-crowd-3.0.1.tar.gz').with({
+          :source => 'https://www.atlassian.com/software/crowd/downloads/binary/atlassian-crowd-3.0.1.tar.gz',
         })}
 
-        it { is_expected.to contain_staging__extract('atlassian-crowd-2.11.1.tar.gz').with({
-          :target => '/opt/crowd/atlassian-crowd-2.11.1-standalone',
-          :creates => '/opt/crowd/atlassian-crowd-2.11.1-standalone/apache-tomcat',
+        it { is_expected.to contain_staging__extract('atlassian-crowd-3.0.1.tar.gz').with({
+          :target => '/opt/crowd/atlassian-crowd-3.0.1-standalone',
+          :creates => '/opt/crowd/atlassian-crowd-3.0.1-standalone/apache-tomcat',
         })}
       end
 
@@ -140,8 +140,8 @@ describe 'crowd::install' do
             download_url => 'http://192.168.0.0.1:8080/',
           }"
         end
-        it { is_expected.to contain_staging__file('atlassian-crowd-2.11.1.tar.gz').with({
-          :source => 'http://192.168.0.0.1:8080/atlassian-crowd-2.11.1.tar.gz',
+        it { is_expected.to contain_staging__file('atlassian-crowd-3.0.1.tar.gz').with({
+          :source => 'http://192.168.0.0.1:8080/atlassian-crowd-3.0.1.tar.gz',
         })}
       end
 
@@ -151,8 +151,8 @@ describe 'crowd::install' do
             download_url => 'http://192.168.0.0.1:8080/',
           }"
         end
-        it { is_expected.to contain_staging__file('atlassian-crowd-2.11.1.tar.gz').with({
-          :source => 'http://192.168.0.0.1:8080/atlassian-crowd-2.11.1.tar.gz',
+        it { is_expected.to contain_staging__file('atlassian-crowd-3.0.1.tar.gz').with({
+          :source => 'http://192.168.0.0.1:8080/atlassian-crowd-3.0.1.tar.gz',
         })}
       end